requirements.txt
streamlit==1.17.0
mysql-connector-python==8.0.32
